refactor(ui/settings): 抽取跟踪文件夹设置到独立文件
将原CustomSettings.ts中的跟踪文件夹设置相关代码抽离到TrackedFoldersSetting.ts，重构设置页面代码结构，提升可维护性
